factor tests out for partition pruning
